将算法文件夹删除master字样，统一命名
