+++ MiDaS/models/base_model.py
